Deprecate github-cli helper
