+++ lazy_github/ui/screens/primary.py
@@ -16,9 +16,10 @@ from textual.widgets import TabbedContent, Tabs
 from lazy_github.lib.bindings import LazyGithubBindings
 from lazy_github.lib.constants import NOTIFICATION_REFRESH_INTERVAL
 from lazy_github.lib.context import LazyGithubContext
+from lazy_github.lib.github.auth import is_logged_in_to_cli
 from lazy_github.lib.github.issues import list_issues
+from lazy_github.lib.github.notifications import unread_notification_count
 from lazy_github.lib.github.pull_requests import get_full_pull_request
-from lazy_github.lib.github_cli import is_logged_in, unread_notification_count
 from lazy_github.lib.logging import lg
 from lazy_github.lib.messages import (
     IssuesAndPullRequestsFetched,
@@ -375,7 +376,7 @@ class LazyGithubMainScreen(Screen):
     async def refresh_notification_count(self) -> None:
         widget = self.query_one("#unread_notifications", UnreadNotifications)
         if LazyGithubContext.config.notifications.enabled:
-            if not await is_logged_in():
+            if not await is_logged_in_to_cli():
                 error_message = "Cannot load notifications - please login to the gh CLI: gh auth login"
                 self.notify(error_message, title="Failed to Load Notifiations", severity="error")
                 lg.error(error_message)
